oboe: move flowgraph from OboeTester to inside Oboe
This will be used internally for doing data conversion within a stream.
This flowgraph API is experimental and is likely to change.
