+++ src/flowgraph/RampLinear.cpp
